🔄 synced local 'skyvern/' with remote 'skyvern/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Extend auto-completion logic to handle emerging HTML elements in `auto-completion-choose-option.j2` and `handler.py`.
>
>   - **Behavior**:
>     - Extend auto-completion logic in `auto-completion-choose-option.j2` to include IDs for emerging HTML elements.
>     - Update `choose_auto_completion_dropdown()` and `input_or_auto_complete_input()` in `handler.py` to handle new elements detected after input actions.
>   - **Misc**:
>     - Add `scraped_page` parameter to `handle_input_text_action()` and `choose_auto_completion_dropdown()` in `handler.py` to facilitate new element detection.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for 8a18c6197437c0245ab010aa5adff2de916c00e7. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->

+++ skyvern/forge/sdk/db/polls.py
@@ -13,7 +13,7 @@ async def wait_on_persistent_browser_address(
     session_id: str,
     organization_id: str,
     timeout: int = 600,
-    poll_interval: int = 2,
+    poll_interval: float = 2,
 ) -> str | None:
     persistent_browser_session = await await_browser_session(db, session_id, organization_id, timeout, poll_interval)
     return persistent_browser_session.browser_address if persistent_browser_session else None
@@ -24,7 +24,7 @@ async def await_browser_session(
     session_id: str,
     organization_id: str,
     timeout: int = 600,
-    poll_interval: int = 2,
+    poll_interval: float = 2,
 ) -> PersistentBrowserSession | None:
     try:
         async with asyncio.timeout(timeout):
